+++ chessboard/pieces.py
@@ -155,7 +155,7 @@ class King(Piece):
     def movements(self):
         """ King moves one square in any direction.
 
-        Don't mind out of bounds relative positions: forbidden one will be
+        Don't mind out-of-bounds relative positions: forbidden ones will be
         silently discarded within the ``Piece.territory()`` method above.
         """
         return set([
@@ -199,7 +199,20 @@ class Bishop(Piece):
 class Knight(Piece):
     """ Knight model. """
 
-    #@property
-    #def movements(self):
-    #    """ Knight moves unrestricted diagonally. """
-    #    return self.diagonals
+    @property
+    def movements(self):
+        """ Knight moves in L shapes in all 8 directions.
+
+        Don't mind out-of-bounds relative positions: forbidden ones will be
+        silently discarded within the ``Piece.territory()`` method above.
+        """
+        return set([
+            # Top-right movements.
+            (+2, +1), (+1, +2),
+            # Top-left movements.
+            (-2, +1), (-1, +2),
+            # Bottom-right movements.
+            (+2, -1), (+1, -2),
+            # Bottom-left movements.
+            (-2, -1), (-1, -2),
+        ])